MP-26: Реализован класс SmartThermostat.java, со всеми заданными параметрами в задании. Написаны тесты.

+++ src/main/java/ru/mentee/power/conditions/SmartThermostat.java
@@ -22,7 +22,7 @@ public class SmartThermostat {
         if (!WEEKDAYS.contains(dayOfWeek) && !WEEKENDS.contains(dayOfWeek)) {
             return ERROR_TEMP_CODE;
         }
-        int targetTemperature = 0;
+        double targetTemperature = 0;
 
         if (WEEKDAYS.contains(dayOfWeek)) {
             if (timeOfDay >= 6 && timeOfDay <= 8) {
